# Backend/openapi2-functions.yaml
swagger: '2.0'
info:
  title: Microservices API Gateway
  description: Gateway to route to microservices
  version: '1.0'
host: 'api.mydomain.com'
schemes:
  - https
basePath: /api/v1
paths:
...
  /start-training:
    post:
      summary: Start a training
      operationId: startTraining
      requestBody:
        description: Training session ID to start
        required: true
        content:
          application/json:
            schema:
              $ref: '#/definitions/StartTraining'
      x-google-backend:
        address: https://TRAINING_MANAGEMENT_COMMANDS_URL/start-training
      responses:
        '200':
          description: Training started
  /stop-training:
    post:
      summary: Stop a training
      operationId: stopTraining
      requestBody:
        description: Training session ID to stop
        required: true
        content:
          application/json:
            schema:
              $ref: '#/definitions/StopTraining'
      x-google-backend:
        address: https://TRAINING_MANAGEMENT_COMMANDS_URL/stop-training
      responses:
        '200':
          description: Training stopped
...
definitions:
...
  StartTraining:
    type: object
    required:
      - user_id
      - training_type
    properties:
      user_id:
        type: integer
      training_type:
        type: string
        description: Type of training session
  StopTraining:
    type: object
    required:
      - session_id
      - end_time
      - duration
      - calories_burned
    properties:
      session_id:
        type: string
        format: uuid
      end_time:
        type: string
        format: date-time
      duration:
        type: integer
        description: Duration of the training session in minutes
      calories_burned:
        type: number
        description: Estimated calories burned during the session
...